pyproject.toml
[tool.poetry]
name = "postgres-multi-agent-data-analytics-base"
version = "0.1.0"
description = ""
authors = ["Epiphane EGAH <egahepiphane@gmail.com>"]
readme = "README.md"
# add packages
packages = [
    { include = "src", from = '.'}
]
[tool.poetry.dependencies]
python = ">=3.10,<3.12"
openai = "0.28"
argparse = "^1.4.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pyautogen = "^0.1.7"
transformers = "^4.38.1"
fuzzywuzzy = "^0.18.0"
scikit-learn = "^1.4.1.post1"
torch = "^2.2.1"
tiktoken = "^0.6.0"
fastapi = "0.85.0"
uvicorn = "^0.27.1"
python-decouple = "^3.8"
gunicorn = "^21.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# add entry point
[tool.poetry.scripts]
start = "entry_point:main"